requirements.txt
plyer~=2.0.0
pushbullet.py~=0.9.1
opencv-python~=4.5.1.48
mysql-connector-python~=8.0.24